refactor(review): share function-local import traversal
